requirements.txt
neo4j==5.11.0
pinecone-client==2.2.1
openai==1.10.0
python-dotenv==1.0.0
networkx==3.2
matplotlib==3.8.1
tqdm==4.66.1
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.11
requests==2.31.0
